LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-wm-expected.txt


FAIL  horizontal-tb ltr assert_equals: offsetLeft expected 10 but got 99999
FAIL flip-inline horizontal-tb ltr assert_equals: offsetLeft expected 360 but got 99999
FAIL flip-inline vertical-lr ltr assert_equals: offsetLeft expected 10 but got 99999
FAIL flip-start horizontal-tb ltr assert_equals: offsetLeft expected 20 but got 99999
FAIL flip-start horizontal-tb rtl assert_equals: offsetLeft expected 340 but got 99999
